+++ src/exams/exam-periods.controller.ts
@@ -6,8 +6,13 @@ import {
   Patch,
   Param,
   Delete,
+  UseGuards,
+  Req,
+  ForbiddenException,
 } from "@nestjs/common";
-import { ApiTags } from "@nestjs/swagger";
+import { ApiBearerAuth, ApiTags } from "@nestjs/swagger";
+import { AccessTokenGuard } from "src/common/guards/access-token.guard";
+import { Action, CaslAbilityFactory } from "src/casl/casl-ability.factory";
 import { CreateExamPeriodDto } from "./dto/create-exam-period.dto";
 import { UpdateExamPeriodDto } from "./dto/update-exam-period.dto";
 import { ExamPeriodsService } from "./exam-periods.service";
@@ -15,10 +20,23 @@ import { ExamPeriodsService } from "./exam-periods.service";
 @Controller("exam-periods")
 @ApiTags("exam-periods")
 export class ExamPeriodsController {
-  constructor(private readonly examPeriodsService: ExamPeriodsService) {}
+  constructor(
+    private readonly examPeriodsService: ExamPeriodsService,
+    private readonly caslAbilityFactory: CaslAbilityFactory,
+  ) {}
 
   @Post()
-  async create(@Body() createExamPeriodDto: CreateExamPeriodDto) {
+  @UseGuards(AccessTokenGuard)
+  @ApiBearerAuth()
+  async create(
+    @Body() createExamPeriodDto: CreateExamPeriodDto,
+    @Req() request: any,
+  ) {
+    const ability = this.caslAbilityFactory.createForExamPeriod(request.user);
+    if (ability.cannot(Action.Create, "all")) {
+      throw new ForbiddenException("Can't create exam period");
+    }
+
     const examPeriod = await this.examPeriodsService.create(
       createExamPeriodDto,
     );
@@ -34,10 +52,18 @@ export class ExamPeriodsController {
   }
 
   @Patch(":id")
+  @UseGuards(AccessTokenGuard)
+  @ApiBearerAuth()
   async update(
     @Param("id") id: number,
     @Body() updateExamPeriodDto: UpdateExamPeriodDto,
+    @Req() request: any,
   ) {
+    const ability = this.caslAbilityFactory.createForExamPeriod(request.user);
+    if (ability.cannot(Action.Update, "all")) {
+      throw new ForbiddenException("Can't update exam period");
+    }
+
     const examPeriod = await this.examPeriodsService.update(
       +id,
       updateExamPeriodDto,
@@ -47,7 +73,14 @@ export class ExamPeriodsController {
   }
 
   @Delete(":id")
-  async remove(@Param("id") id: number) {
+  @UseGuards(AccessTokenGuard)
+  @ApiBearerAuth()
+  async remove(@Param("id") id: number, @Req() request: any) {
+    const ability = this.caslAbilityFactory.createForExamPeriod(request.user);
+    if (ability.cannot(Action.Delete, "all")) {
+      throw new ForbiddenException("Can't delete exam period");
+    }
+
     const examPeriod = await this.examPeriodsService.remove(+id);
 
     return this.examPeriodsService.mapToDto(examPeriod);
